Disable delay signing for Desktop tests
To run the desktop test we need to have the test assembly get signed with the
test.snk. Currently we have some workaround in some of the test projects to
disable the delay signing. The change here is to centerline this in the tools
and avoid duplicating the workaround in every test project