Ensure Desktop assets work with packages.config
Desktop projects by default use packages.config, we shouldn't require
folks to use project.json + RID if we don't actually depend on RID
selection as is the case for most desktop packages.

We should make sure that project types that can use packages.config
are supported by our packages.  I've added empty RID entires to the
DefaultValidateFramework list for all of these frameworks.

I'm also addressing a problem where we might get a different asset when
someone uses a RID over not having a RID: eg the package has lib/net45
asset that is different from its runtimes/win/lib/netstandard1.1 asset.
To catch this case I added a check that tries to see if there is a
better matching asset by framework for any runtime asset.

This caught a number of cases where we were making this mistake for
desktop, so I've made a general rule that any desktop project with a
RID will also put the asset in the RID-less folder.