AddCloudTestContainerSuffix,  a new property optionally appended to the Azure Container name
This allows differentiating between different target OS'es.  Previously TargetOS was used, but now that we need this value to be set properly and since when set properly it can be the same across different OS'es (I.e. all Windows SKUs get Windows_NT) we need a mechanism to keep the containers separate.