Don't continue on error
Previously I set continue on error for a number of the packaging tasks.
The reasoning was that it is often helpful to let the process complete
and examine the nuspec to visualize what caused the error.  The problem
with this is that ContinueOnError does not cause MSBuild to return non-
zero when being run on multiple nodes.  As a result this was letting
folks pass CI even with packaging errors.

Remove usages of continue on error.