Support CSProj > PkgProj references
This change adds support for CSProjs to reference PkgProjs.
The PkgProj will be built and then all files that compose it along
with their target path in the package will be provided to the
referencing project.  The referencing project will then evaluate
those target paths using the TFM/RID as appropriate to determine
which assets to reference and copy to the output directory.

This can be used by our tests to simplify project references to
product binaries.