Change default to not use slash for qualifiers
When tools need to parse arguments for other platforms (i.e. Linux/OSX)
they have files that begin with "/" and that messes up the parser and
causes an "undefined qualifer" error. The best way to handle this is
to disable the uses of slash for qualifiers by default. People can still
manually enable it if they chose to want to support it but for most
of our tools that are now running x-plat we need to disable this and
only support dash for qualifiers.