Rearrange targets into groups and add comments
This change is just re-ordering targets to more closely match their
execution order and adding some comments around the groups of
targets.

I did this to help with the previous change but wanted to commit in
an separate diff so that one diff has funcitonal changes and the other
is purely moves & comments.