Automatically ref RID-graph & lineup packages
Required for https://github.com/dotnet/corefx/issues/8239

##Fat packages
A dependency is added on the RID-graph package for any package
that uses `runtimes`.  The dependency is constrained to only those
frameworks that may need the RID-specific lib.

##Split packages
A dependency is added on the lineup and RID-graph package for any that
has runtime packages and doesn't include a runtime.json itself.  The
dependency is constrained to only those frameworks that are not
satisfied by the ref-package exclusively.