Switch test framework to netcoreapp1.0
This switches the default framework and the test framework from
DNXCore50 to netcoreapp1.0.  I've also taken the opportunity to clean
up all non-essential refs to DNXCore50, DNXCore, and DNX.

The only remaining references to these deprecated monikers are:
1. Import statements: needed until all packages are updated.
2. Microsoft.DotNet.Build.Tasks.Packaging.Tests: needed until we can
update the version of BuildTools used by BuildTools to get this change.