Add the ability to filter out RIDs when building tests against packages. (#760)
We have a situation where new RIDs have been added to the product but we
don't yet have the actual packages; as a result the test build fails when
buliding against packages as it can't resolve these runtime packages. To
work around this I have added an optional property to filter out RIDs when
generating the test project.json files.