Fix for building test projects in VS and rebuildandtest on command line
This is a fix for corefx to allow test projects that now reference the pkgprojs to build in VS.
MS.Common.Targets hardcodes GetTargetPath when resolving projectrefs in the IDE.
Currently VS doesn't understand the p2p reference test-runtime, so the restore must be done via msbuild.

This also resolves rebuildandtest on commandline msbuild, by mapping it to the first usage of ProjectReference
in a target which happens on both the Build and Rebuild paths.

Changes done with @ericstj 's help