Merge pull request #672 from maririos/parser - The "run" tool - dev workflow
The run tool has a published contract of inputs and outputs to encapsulate the work we've been doing related to the dev workflow. It parses arguments and formats them in order to run the command given by the user. It also provides documentation of the most common settings we use per repo, and the commands we can execute.
Each command has a set of options/alias that will trigger a different behavior.