EnsureOOBFramework should promote impl deps
EnsureOOBFramework makes sure that an oob implementation will always
take precedence over a placeholder in the case we have a framework that
transitioned from targeting-pack to package as was the case with
netcore50.

The problem here is that if it promotes only one of the ref or lib we'll
only get the dependencies from that asset, since the other asset's
dependencies will be in the less specific dependency group.

This became a problem when we started merging runtime specific
assets into the same packages as the ref: the fat package work.
Since a netstandard lib that is rid specific will be preferred over the
placeholder that is not rid specific we didn't need to copy the
implementation.   The result was the dependencies would only
be in the netstandard group.

To fix this I added a phase to find the actual implementation and copy
it to a 'netcore50' framework item.  I left a marker in the targetpath
of the copied items so that we wouldn't package the duplicate.

Longer term I'd like to rewrite how we do all the dependency harvesting
so that it is less brittle.  I think the right way to do this would be
to do actuall asset selection for TFM + TFM/RID tuples as appropriate
for all supported app models, get the actual assets that nuget will
select and harvest the dependencies at that point.

Port of 2c958d3472ec3fb925c9d142c4c91bd4731a780c.