Downgrade error when RID set on ref assm
Previously I'd force an error if a reference assembly was ever
accidentally packaged as a RID-specific asset.  I've seen a couple cases
recently where folks are running into this when specifying configuration
parameters and custom SLN configurations.

Rather than block in these cases, I'll just clear the property and let
the build complete.  If this actually causes a real problem we'll catch
it as a binclash (since ref's never binplace to RID specific paths).