Add RepoUtil sources to the build tools repo.
We have a tool, RepoUtil, created by the Roslyn team for implementing the
produces and consumes repo APIs.  Since this is meant to be a general
purpose tool to be consumed by various repos I'm moving the sources to the
build tools repo.  I have also added nuget package authoring for this tool
so that it can easily be consumed by other repos.  In the process I have
converted the tool to be a .NET Core app.