Refactor package destination calculation
This change refactors how we calculate the destination path of a library
in a package.

Previously PackageTargetFramework property (single value) or
PackageDestination (many values, precalculated path) must be
specified.

This had a few problems:
1. PackageTargetFramework was often the same information that was in
NuGetTargetMoniker.
2. The only means for specifying more than one destination was too
verbose and required duplicating the target moniker in the path and
TargetFramework metadata.

This fixes both of those issues by allowing folks to specify a list of
PackageTargetFrameworks either in the PackageTargetFramework
property or item.  When using the item each PackageTargetFramework
can optionally specify a TargetRuntime as well.  Additionally, if folks
wish to package using the same target moniker as NuGetTargetMoniker,
we will fallback to using this if nothing else is specified.

In doing this I refactored how we do our asset promotion for ref/lib for
desktop to make it clearer and more consistently named.  This is a minor
breaking change, so I mitigated that by preserving the old property name
PackageAsRefAndLib.

This change also removes the ValidatePackageTargetFramework target.
That target's purpose was to calculate the netstandard version of an
assembly based on references but its no longer needed since we can now
use NuGet.  It's replacement is validation in GetPackageDestination
which ensures that we package things using a compatible TFM to what
we build them against.