Place the run tool into its own nuget package. (#987)
To allow repos to consume the run tool without pulling down all of build
tools. For now continue to place the run tool in the build tools package
until we're ready to make the complete split.
Include specific assemblies in the RunTool nuget package that aren't a
part of the shared framework.