Add PreRelease support to PackageIndex
This allows for PackageIndexes to specify a PreRelease.  This can
be specified at the index level or package level in the package index
file.

When applying a PreRelease string to a package version,
ApplyPreReleaseSuffix will prefer the one from the index if present,
then fallback to using the repo's build string.