Make package harvesting replace refs
Previously we'd always prefer live assets over harvested assets from the
previous shipped package.  That put the burden on the developer to
know when to include references.

This change lets us keep references in the package and only use them
when they are significant.  This should help reduce the concept count
required by folks working in corefx.