Remove EnsureOOBFramework
This target was responsible for automatically finding the best asset to
use for NETCore50.  It needed to determine if any portable (netstandard)
asset would work then copy that to the netcore50 folder so that it would
take precedence over a placeholder file.

This has become problematic since we now may have netcore50 assets
coming from a harvested package or from the live build.  If we run
EnsureOOBFramework before harvesting then we can never remove
netcore50-specific project configurations from the build since removing
those would cause this to prefer a live built netstandard asset.  If we
run this after harvesting we'd always prefer the harvested netcore50
asset over a copied live-built netstandard asset.

The solution is to just be explicit in the projects about which live
assets should be used for netcore50.  That eliminates the need for this
target and removes the "magic" from the packaging.