Make PromoteDependencies copy lib>ref & ref > lib
Previously we only handled cases where the lib TFM was more specific
than ref, so we copied the ref dependencies to the lib TFM group.

The same can happen the other way around, for instance when we build
a single netstandard lib, expose all API as netcoreapp, but only some
as netstandard.  Also when we expose the ref as netcore50, but the lib
as netstandard.