GenAPI: Ignore obsolete-as-error ctors when generating base call
The first constructor in the base class is sometimes one that is marked
as [Obsolete("message", true)] meaning any usage of that ctor results
in a compilation error. Example: System.ComponentModel.AsyncCompletedEventArgs

We now skip those constructors when generating the base call so classes
that inherit from those types compile out of the box.