Fix default test runs from VS
Change
https://github.com/dotnet/buildtools/commit/499458e3889a68b633be1bccc31eae6fe4f30191
broken running tests in VS when you didn't actually build from the command line first.

This change hooks the test publishing step, not the running, to the PrepareForRun
so that the default test configuration can be F5'ed and work in VS.

Note this isn't full-proof because it does rely on the default configuration to
exist and be .NET Core based but it will enable most of our test projects to just work.