Allow placeholders from pkgproj refs
This allows placeholders from a pkgproj ref.  Previously we'd error
when a pkgproj reference provided only placeholders.  The idea was
"why even reference the project if you don't get anything out of it?"
but this makes constructing the project references cumbersome:
consider a project that builds for netstandard1.3 but needs to run on
net46.  No implementation may be needed for net46 since many contracts
are inbox, but we don't want to force folks to condition their project
references on the TestTFM.