Fix test script generation
RunTestsForProject not only runs the tests but also generates the
test runner script. As part of a recent clean-up that aspect was missed
with change https://github.com/dotnet/buildtools/commit/499458e3889a68b633be1bccc31eae6fe4f30191.

To get our official test runs working again we need to move the condition
to of TestDisabled!=true to the actual execution of thests instead of the
RunTestsForProject target.

In the future we need to split this logic into mutliple targets if we need the
logic for different purposes.