Do not redirect std[out|err] when running command
Previously, we would redirect both the standard output and error streams
when run.exe would run a child process. This allowed us to "batch up" all
the output to stderr and report it at the end of the command.

The redirection prevented colorized output from working (on Win32 the
console color commands are no-ops when the output streams are no going to
a console, and similar logic on Unix causes ANSI escape sequences for
colors to not be included when the output is redirected).

This was a significant regression for MSBuild which uses colors. The
buffering we do for standard output/error also prevented tools from
interleaving errors/warnings and regular output.

Stop doing this redirection. Child processes which spew output to stderr
and then doing "replay" it at the end of command execution will no longer
have their output at the end of a run invocation. However, we get back
colors and child command execution works the same way now as if you had
invoked the command directly.

Fixes #864