Fix symbol package generation on Unix
We were using a windows path seperator when setting the TargetPath
metadata on source files we added to the ItemGroup of artifacts to pack.
This was causing the logic that splits the nuget package into a product
and symbols package to place the source assets in the product package in
an oddly named 'src\src" directory.

The NuGet logic around dealing with path seperators cross platform has
known issues (e.g. NuGet/Home#3584).

For now, we will use the correct path seperator when adding the metadata.
This seems to be the cleanest solution (I considered changing the path
seperators in the GenerateNuSpec target, but I worry about doing path
normalization there).

Contributes to dotnet/corefx#12758.  This actually fixes the issue, but of
course they will need to tae buildtools update.