Ensure we're comparing 3-part versions
NuGetVersion was returning a 4-part version even though the version
string it was constructed with only had 3 parts.  As a result, when we
compared it to the 3 part version in the stable version list it came out
as greater even though it was the same.  Fix this by ensuring we are
comparing 3-part versions.