Update to run on Jenkins
I have ripped out most of the guts that were in the PerfTesting.targets
file and moved everything over to run using the dotnet cli and the new
dev/eng model.  This includes removing all of the previous goo that
helped with loading the analysis and runner exes as those have been
replaced by the PerfRunner.