Don't create layout for runtime packages
The identity package has all the layout information for all runtime
packages, so we shouldn't try to copy the layouts for both id and
runtime packages.  When we do it would cause a race if the two
were building at the same time.

I was trying to avoid this but didn't have the condition correct.