Uptake the latest buildtools package, and new CLI (#1294)
* Uptake the latest buildtools package, and new CLI

The buildtools package uses shared framework 1.1.0 instead of 1.0.0, so switch to a matching CLI package. This is the same CLI version as CoreFX and CoreCLR use.

* Fix packaging tests: BuildTools TestPath change

Now more is expected in TestPath that determines the test root. TFM is not added by testing targets anymore. See CoreFX PR 12634.

* Upgrade Run tool, ensure ConfigurationGroup passed

Without the entry in defaultValues, no ConfigurationGroup was passed to the msbuild command. This caused bin clashes due to a projects not having ConfigurationGroup passed, but the ProjectReferences did have it set due to dir.props.