Ensure PackageIndex is not mutated during build
Previously we'd merge any index we saw during build into an index
singleton.  This was because initially the index was seen as repo data
and not project specific data.  I broke that assumption when I started
using a different index in packages which wanted a different baseline.
I began using indexes specific to individual projects.  This caused
state to bleed across packages.

Fix this by always calling Load and backing it with a cache that will
not reload the index if the same index set was already loaded.