Move inbox data to the PackageIndex
This removes the use of framework lists to define which packages are
inbox.  Instead we'll let that data live with the repro consuming
buildtools and be part of the PackageIndex.  I've added arguments
to UpdatePackageIndex that let folks migrate.  I've done the migration
for corefx and tested that this behaves as expected.

While doing this I also made some changes to make the index more
readable by sorting various things (package list, version list, etc).