Get and use IBCMerge tooling from internal/authed VSTS feed (#1322)
* Get and use IBCMerge tooling from internal feed

This adds "internal-tool-runtime" restore just before Sync for tooling that needs to come from an authed NuGet feed. On Windows with full framework, a NuGet.Config is created on disk with an encrypted password. On *nix with Core, the unimplemented approach is to pipe a in-memory NuGet.Config into the restore command that has the credentials in plaintext. (The type of user encryption used on Windows isn't supported xplat, and the piping workaround doesn't work on Windows.) The hope in the future is to implement both methods as part of the bootstrap flow, once CoreFX uses it.

IBCMerge is the first tool that we need to restore this way. I changed the optimization targets to pull it in using the new internal tool restore flow. I made some changes to support the new file type we're using to move IBC data into the build. The IBC data package id, version, and source are all configurable by the repository using the optimization targets, so auto-upgrade can happen independent of buildtools.

* Review: remove pointless null/empty arg echo