Make ValidatePackage use index inbox algorithm
Previously we were duplicating the inbox algorithm with slight
differences between ValidatePackage and the index.

The previous algorithm was missing all compatible frameworks since
those wouldn't have explicit entries in the new index data model.

The change uncovered a latent issue in the WCF packages where we've
shipped contracts with higher versions than what was inbox for the same
API version.  To permit this I've added a suppression flag, which can
be used by packages as follows.

```
<ItemGroup>
  <ValidatePackageSuppression Include="PermitInboxRevsion" />
</ItemGroup>
```