Make harvesting explicit
Previously harvesting would happen automatically based on what was found
in the package index, harvesting all applicable paths and only allowing
for opt-out.  We'd harvest both the supported frameworks from the
previous package as well as all the files.

Now the files will only be harvested if specified via the
HarvestIncludePaths item.  Supported frameworks will still be harvested
regardless of this item since we need to ensure that new packages
support the same frameworks as old.

The old harvest all with opt-out behavior can be turned on if desired by
setting HarvestIncludeAllPaths.