Validate implementation for netstandard
Previously we didn't validate if netstandard references had a matching
implementation.  Instead we relied on guardrails to catch that and only
validated for the frameworks we cared about and wanted guardrails to
succeed.

We are no longer relying on guardrails nor allowing netstandard refs
without a matching implementation.  Validate for this, and add a
suppression for cases where we explicitly want to break this rule.