Enable build of partial reference facades
Adds a mode to GenFacades that starts with an assembly and replaces any
types present in the seeds with typeforwards.  Types not in the seeds
will remain in the assembly.

We'll use this for generating desktop reference facades.

To use this feature set IsPartialFacadeAssembly=true in a reference
project.

After the source is compiled GenFacades will examine all references and
replace any type def in the reference assembly that exists in the
references.

To keep a type in the reference assembly even if it exists in a
reference create an item in the reference project as follows:
```
    <SeedTypePreference Include="namespace.type">
      <Assembly>assemblyName</Assembly>
    </SeedTypePreference>
```