We need to change the casing for the CSharp targets import
Some of the portable targets we are importing have the case as "Targets"
so we have to either stop using those or fix them in order to remove
this and the rename hack in init-tools.sh. Until then we need both
of these in order to have a consistent set of file casing across the
different imports consumers of build tools use.