Do not publish standalone when deploying the tool-runtime
We already tried to run all of our applications against the shared
framework anyway, but we continued to publish the tool runtime as a
standalone application. This bloats the size of the Tools folder.

In addition, this was causing problems when I tried to move the .NET
CLI version we were using forward. In some cases it looked like we
were not actually using the runtime from the shared framework folder.

Since we are on a run on the shared framework runtime plan, we should
stop publishing standalone. This also reduces the amount of code that
crossgen.sh needs to process (since it doesn't need to crossgen
anything in the shared framework).