Enable XUnit conditional filtering uses other assemblies
Conditional Fact and Theory in our xunit filtering is always used types from the test assembly that currently running. We are working on moving PlatformDetection code to CoreFx.Private.TestUtilities assembly as it is naturally fit there and avoid including the cs files in every test project need to use it. We have many tests using Conditional Fact/Theory which calls PlatformDetection methods. The way we apply this filtering is we create the PlatformDetection type from the running test assembly. After we move PlatformDetection to CoreFx.Private.TestUtilities, this mechanism will not work. To make it work, we adding the way to allow specify the assembly containing the type through providing the qualified type name which include the assembly name.