Enable CoreFX build to tag DLLs with GITHub URL
In PR #1606 we added logic so that the version resource had not just the GIT commit ID
but also the full Github URL.   This worked well for native code but it turns out that
It did not work for managed code.

This was mostly because for managed code the variables needed to be 'exported' explicitly to a *.props
file so that another version of MSBUILD would pick it up.   There was already a variable that had this
behavior called BuiltByString, and so this fix is to simply put the additional information for the GIT commit
ID into this variable (which would then be attached appropriately.

While I was at it I change the syntax a bit to make it more friendly to maching parsing.   key-value pairs
now always begin with an @ and are separted by spaces (e.g.  @BuiltBy: vancem @SrcCode: http//github.