Allow generating resource with assembly names have dashes
As we generate cs code for accessing the embedded resources. The generated type will be named as FxResources.AsemblyName.SR. if the AssemblyName has characters not allowed in the namespace names (e.g. dashes), this will cause compilation errors. The fix here is we normalize the assembly name before we use it.