Fix race condition because of assembly name normalization (#1653)
* Fix race condition because of assembly name normalization

We used to generate the normalized assembly name in the target GenerateResourcesSource and we override the resw name of the resource file there too. In the incremental build the target GenerateResourcesSource will get skipped but overriding the resw item name still be done as it is used outside the target in other places. This will cause the normalized name property not initialized but it is used in setting the resw name. and that will cause a race condition because many projects will try to access the exact same resw file name while building.
The fix is to move the assembly name normalization outside GenerateResourcesSource so we’ll be sure we always initialize the normalized assembly name property.

* Update the Target name

* Make the task inherit from BuildTask