Ensure deps file contains correct extension
Previously the GenerateDepsJson task was assuming all assemblies were
dlls, however that is not the case; xunit.console.netcore.exe is an exe.

This didn't seem to cause an issue if the file existed in the
application folder.  If xunit.console.netcore.exe existed in the app folder, the host would
ignore the incorrect entry in the framework deps file.

If, however, the file was missing from the app folder and had incorrect
extension, the host would fail at startup.

This change fixes the deps file to have the correct extension.