Make qualifier expansion explicit and eliminate qualifier policies
Previously we had applied Qualifiers to a RID by copying that RID's imports,
applying the Qualifier to all, then importing the previous RID.  This created very
unusual compatibility precedence.

For example, consider `win8-x86: [ win8, win7-x86 ]`

When applying `aot` we'd get `win8-x86-aot : [ win8-aot, win7-x86-aot, win8-x86 ]`

This can cause a `win7-x86-aot` asset to be preferred over `win8-x86` asset, however a
`win-x86-aot` asset would not be preferred over a `win8` asset.

What was lacking from this is a clear precedence relation of the components.
EG: First prefer something that's an exact match, then prefer matching
version and qualifier, then prefer matching version and architecture, then just version,
then previous version (with others matching).

I first tried to solve this with VersionPrecedence QualifierPolicy but that was a poor model,
because I was operating on strings where I had lost the context of component parts.

Instead I've switched to a fully expressed set of precedences.  This isn't the most general
but I think it provides the best clarity.  For now this represents a single precedence:
Version, qualifier, architecture

If in the future we need more precedences we'll have to refactor this, one thing to watch
out for in particular is that if version doesn't have the highest precedence we'll have to
flatten the version list to imports for a single RID for compatible versioned groups.  That
will not scale well which is another reason I've avoided exposing such functionality.

Using the CompatibilityMap is particularly helpful for examining the final precedence
and seeing if it agrees with a component precedence model.