Fix the tool-runtime restore warnings
Moved BuildTools back to a version of CCI which depended on all stable
package versions. This version happens to align with .NET Core 1.0 but
we don't currently have a newer version of that package with stable
dependencies.

Disable NU1701 that we are getting for Microsoft.Net.Compilers.Targets.NetCore
as we know it is safe to resolve its assets as net461. We also plan to remove
this package when we start compiling against the CLI itself instead of the
targets/msbuild we carry with BuildTools.

Delete tool-runtime\project.json because it is no longer used after
https://github.com/dotnet/buildtools/pull/1743