Fix failure on Unix associate with convering PDBs
Code to convert portable pdbs to windows pdbs was being run on UNIX machines.
This is clearly not needed and was causing failures.   Fix it to only run on non Windows
and only for coverage (which is the only thing we know that needs the Windows  PDBs).

See  https://github.com/dotnet/corefx/issues/24807.