Fix race associted with SourceLink
The sourcelink file is generated one for the whole repo so it is possible that
differnet parts of the build will race to write it.   We have seen this show up as a file
access error in builds.

It is already the case that there is logic that avoids running the code most of the time
(see VersionPropsImported).  but races are still possible.  They are avoided in that
scheme by testing for the file existing and simply warning if there is a error trying to
write the file.   We follow suit here.

As a unrelated change there was already a variable SourceLinkFilePath, for the source
link file so I use it instead of building it out of more primitive variable.