Fix to unbreak Linux build.
This is a second attempt.   The first attempt was a misunderstanding of how Linux builds worked.

Before trying to support protable PDBs, GenFacades had a condition that it was never called if
Portable PDBs are used.   However we WANTED to used portable PDBs so we added support for
that and dropped the condition to skip GenFacades if portable PDBs were used.

However this had the effect of running GenFacades on Linux, where previously it never ran there
(because Linux always used portable PDBs).

This fix basically puts the behavior back EXACTLY as it was before on non-windows platforms.
This will fix errors (they happen to be index out of range errors) on Linux builds (because we
are running GenFacades where we should not be).