Add ILAsm to buildtools
Windows-only for now. I think we would want the bootstrapper (the thing that invokes init-tools.cmd) to pass us a RID it used to restore .NET CLI so that we can restore the appropriate version of the native tools. RID in the Windows batch file was easy to hardcode since we seem to hardcode x64 on Windows everywhere.