Port resources file changes to 2.0.0 (#1834)
* Add task to get the resources and merge them into the common pri for external dependencies (#1630)

* Add task to get the resources and merge them into the common pri for external dependencies

* PR Feedback

* PR Feedback, remove System.Windows.Forms dependency, and change target Outputs

* Make RewResourceWriter IDisposable

* Fix assembly filtering when it doesn't need resources to be extracted

* Change to NS implementation using MetadataReader instead of Assembly.LoadFrom

* Add option to disable external resources extraction and use runner's default pri file if common pri doesn't exist

* Allow generate resources with assembly name has dashes (#1642)

* Allow generating resource with assembly names have dashes

As we generate cs code for accessing the embedded resources. The generated type will be named as FxResources.AsemblyName.SR. if the AssemblyName has characters not allowed in the namespace names (e.g. dashes), this will cause compilation errors. The fix here is we normalize the assembly name before we use it.

* small delta

* Make ExtractResWResourcesFromAssemblies derive from BuildTask

This is necessary so that the AssemblyResolver unifies dependencies.

* Fix race condition because of assembly name normalization (#1653)

* Fix race condition because of assembly name normalization

We used to generate the normalized assembly name in the target GenerateResourcesSource and we override the resw name of the resource file there too. In the incremental build the target GenerateResourcesSource will get skipped but overriding the resw item name still be done as it is used outside the target in other places. This will cause the normalized name property not initialized but it is used in setting the resw name. and that will cause a race condition because many projects will try to access the exact same resw file name while building.
The fix is to move the assembly name normalization outside GenerateResourcesSource so we’ll be sure we always initialize the normalized assembly name property.

* Update the Target name

* Make the task inherit from BuildTask