Add target reverse APICompat check for libraries
There are a lot of libraries that have a 1:1 ref and implementation
which want to keep the API surface in sync. In order to enforce that
we have a reverse APICompat check that will verify that the ref
has a matching compatibile set of APIs that the implementation has.

This target is called RunMatchingRefApiCompat. It is only enabled
if RunApiCompat=true and there are no ReferenceFromRuntime references.
It is skipped for projects with runtime references because those are
typically many implementation to one ref which doesn't work as well
for verifying the APIs match.

It can be individually controlled per-project:
1) RunMatchingRefApiCompat = false to disable the check completely
2) Add a baseline file MatchingRefApiCompatBaseline[.$(TargetGroup)]txt